Change how stub works, change file locations
