+++ express-api/routes/ltsaRouter.ts
@@ -4,6 +4,6 @@ import express from 'express';
 const router = express.Router();
 
 // Endpoints for LTSA title information
-router.route('/ltsa/land/title').get(controllers.stubController);
+router.route('/ltsa/land/title').get(controllers.getLTSA);
 
 export default router;
