+++ express-api/utilities/stubResponse.ts
@@ -1,0 +1,5 @@
+import { Response } from 'express';
+
+export const stubResponse = (res: Response) => {
+  return res.status(501).send('Not yet implemented');
+};
